workspace-manager/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
python-multipart==0.0.22
sqlalchemy==2.0.46
passlib[argon2]==1.7.4
docker==7.1.0
itsdangerous==2.2.0
pytest==9.0.2

# NEXUSGATE Integration Dependencies
PyJWT>=2.8.0
cryptography>=41.0.0
